Add custom errorCatch JC-550: 1. Add 'errorCatchForComponent' for 'registration' and 'login' component 2. Add 'errorCatchForService' for 'auth.service'
